add bun idleTimeout = 0 and revert "fix stream event idle for 10s, claude code will retry the request with stream=false"
